move reminder from dashboard to header and remove fixed position
