+++ src/components/Dashboard/Dashboard.jsx
@@ -1,8 +1,6 @@
 import { useState, useEffect } from 'react';
-import { Row, Col, Container, Button, Card } from 'reactstrap';
+import { Row, Col, Container } from 'reactstrap';
 import { connect } from 'react-redux';
-import { ENDPOINTS } from 'utils/URL';
-import axios from 'axios';
 import Leaderboard from '../LeaderBoard';
 import WeeklySummary from '../WeeklySummary/WeeklySummary';
 import Badge from '../Badge';
@@ -17,13 +15,6 @@ export function Dashboard(props) {
   const [summaryBarData, setSummaryBarData] = useState(null);
   const { match, auth } = props;
   const displayUserId = match.params.userId || auth.user.userid;
-  const userId = match.params.userId || auth.user.userid;
-  const [isModalVisible, setModalVisible] = useState(false);
-  const [modalContent, setModalContent] = useState('');
-  const [userDashboardProfile, setUserDashboardProfile] = useState(undefined);
-  const [hasProfileLoaded, setHasProfileLoaded] = useState(false);
-  const dismissalKey = `lastDismissed_${userId}`;
-  const [lastDismissed, setLastDismissed] = useState(localStorage.getItem(dismissalKey));
 
   const toggle = () => {
     setPopup(!popup);
@@ -34,95 +25,6 @@ export function Dashboard(props) {
       }
     }, 150);
   };
-
-  const closeModal = () => {
-    setModalVisible(false);
-    const today = new Date();
-    localStorage.setItem(dismissalKey, today);
-    setLastDismissed(today);
-    // console.log('modal dismissed...it will pop up again next thursday...');
-    // console.log('dismmisal key: ', dismissalKey);
-  };
-
-  const getMostRecentThursday = date => {
-    const mostRecentThursday = new Date(date);
-    if (date.getDay() === 4) {
-      // If today is Thursday, return today's date
-      mostRecentThursday.setHours(0, 0, 0, 0);
-      return mostRecentThursday;
-    }
-    // Otherwise, find the previous Thursday
-    mostRecentThursday.setDate(date.getDate() - ((date.getDay() + 3) % 7));
-    mostRecentThursday.setHours(0, 0, 0, 0);
-    return mostRecentThursday;
-  };
-
-  const loadUserDashboardProfile = async () => {
-    if (!userId || hasProfileLoaded) return;
-    try {
-      const response = await axios.get(ENDPOINTS.USER_PROFILE(userId));
-      const newUserProfile = response.data;
-      setUserDashboardProfile(newUserProfile);
-      setHasProfileLoaded(true); // Set flag to true after loading the profile
-    } catch (err) {
-      // console.log('User Profile not loaded.', err);
-    }
-  };
-
-  useEffect(() => {
-    loadUserDashboardProfile();
-
-    if (
-      auth.user.role === 'Owner' ||
-      auth.user.role === 'Administrator' ||
-      auth.user.role === 'Mentor'
-    ) {
-      setModalVisible(false);
-      // console.log('No notification modal...');
-      return;
-    }
-
-    const today = new Date();
-    const lastDismissedDate = lastDismissed ? new Date(lastDismissed) : null;
-
-    // Check if lastDismissed date is ahead of today
-    if (lastDismissedDate > today) {
-      // Clear lastDismissed in both the state and localStorage
-      setLastDismissed(null);
-      localStorage.removeItem(dismissalKey);
-    }
-
-    // Check if today is Thursday or the stored date is before the most recent Thursday
-    if (!lastDismissed || lastDismissedDate < getMostRecentThursday(today)) {
-      if (userDashboardProfile?.teams && userDashboardProfile.teams.length > 0) {
-        // console.log('On a team...Header never shown this week...', getMostRecentThursday(today));
-
-        if (auth.user.role === 'Assistant Manager' || auth.user.role === 'Volunteer') {
-          setModalVisible(true);
-          setModalContent(`If you are seeing this, it’s because you are on a team! As a member of a team, you
-          need to turn in your work 24 hours earlier, i.e. FRIDAY night at midnight Pacific
-          Time. This is so your manager has time to review it and submit and report on your
-          entire team’s work by the usual Saturday night deadline. For any work you plan on
-          completing Saturday, please take pictures as best you can and include it in your
-          summary as if it were already done. By dismissing this notice, you acknowledge you
-          understand and will do this.`); // Assistant Manager or Volunteer message
-        } else if (auth.user.role === 'Manager') {
-          setModalVisible(true);
-          setModalContent(`If you are seeing this, it’s because you are a Manager of a team! Remember to turn 
-          in your team’s work by the Saturday night at midnight (Pacific Time) deadline.
-          Every member of your team gets a notice like this too. Theirs tells them to get you their 
-          work 24 hours early so you have time to review it and submit it. If you have to remind them 
-          repeatedly (4+ times, track it on their Google Doc), they should receive a blue square.
-          `); // Manager message
-        }
-      }
-    } else {
-      setModalVisible(false);
-    }
-
-    // console.log('profile: ', userDashboardProfile);
-    // console.log(auth.user.role);
-  }, [lastDismissed, userId, userDashboardProfile]);
 
   useEffect(() => {
     // eslint-disable-next-line react/destructuring-assignment
@@ -141,68 +43,52 @@ export function Dashboard(props) {
 
   return (
     <Container fluid>
-      <div className="floating-header">
-        {match.params.userId && auth.user.userid !== match.params.userId ? (
-          <PopUpBar component="dashboard" />
-        ) : (
-          ''
-        )}
-        <SummaryBar
-          displayUserId={displayUserId}
-          toggleSubmitForm={toggle}
-          role={auth.user.role}
-          summaryBarData={summaryBarData}
-        />
-        {isModalVisible && (
-          <Card color="primary">
-            <div className="close-button">
-              <Button close onClick={closeModal} />
-            </div>
-            <div className="card-content">{modalContent}</div>
-          </Card>
-        )}
-      </div>
+      {match.params.userId && auth.user.userid !== match.params.userId ? (
+        <PopUpBar component="dashboard" />
+      ) : (
+        ''
+      )}
+      <SummaryBar
+        displayUserId={displayUserId}
+        toggleSubmitForm={toggle}
+        role={auth.user.role}
+        summaryBarData={summaryBarData}
+      />
 
-      <div
-        style={{
-          paddingTop: isModalVisible ? '200px' : '130px',
-        }}
-      >
-        <Row>
-          <Col lg={{ size: 7 }}>&nbsp;</Col>
-          <Col lg={{ size: 5 }}>
-            <div className="row justify-content-center">
-              <div
-                role="button"
-                className="mt-3 mb-5 text-center"
-                onClick={toggle}
-                onKeyDown={toggle}
-                tabIndex="0"
-              >
-                <WeeklySummary isDashboard isPopup={popup} displayUserId={displayUserId} />
+      <Row>
+        <Col lg={{ size: 7 }}>&nbsp;</Col>
+        <Col lg={{ size: 5 }}>
+          <div className="row justify-content-center">
+            <div
+              role="button"
+              className="mt-3 mb-5 text-center"
+              onClick={toggle}
+              onKeyDown={toggle}
+              tabIndex="0"
+            >
+              <WeeklySummary isDashboard isPopup={popup} displayUserId={displayUserId} />
+            </div>
+          </div>
+        </Col>
+      </Row>
+      <Row>
+        <Col lg={{ size: 5 }} className="order-sm-12">
+          <Leaderboard displayUserId={displayUserId} />
+        </Col>
+        <Col lg={{ size: 7 }} className="left-col-dashboard order-sm-1">
+          {popup ? (
+            <div className="my-2">
+              <div id="weeklySum">
+                <WeeklySummary displayUserId={displayUserId} setPopup={setPopup} />
               </div>
             </div>
-          </Col>
-        </Row>
-        <Row>
-          <Col lg={{ size: 5 }} className="order-sm-12">
-            <Leaderboard displayUserId={displayUserId} />
-          </Col>
-          <Col lg={{ size: 7 }} className="left-col-dashboard order-sm-1">
-            {popup ? (
-              <div className="my-2">
-                <div id="weeklySum">
-                  <WeeklySummary displayUserId={displayUserId} setPopup={setPopup} />
-                </div>
-              </div>
-            ) : null}
-            <div className="my-2" id="wsummary">
-              <Timelog isDashboard passSummaryBarData={setSummaryBarData} />
-            </div>
-            <Badge userId={displayUserId} role={auth.user.role} />
-          </Col>
-        </Row>
-      </div>
+          ) : null}
+          <div className="my-2" id="wsummary">
+            <Timelog isDashboard passSummaryBarData={setSummaryBarData} />
+          </div>
+          <Badge userId={displayUserId} role={auth.user.role} />
+        </Col>
+      </Row>
     </Container>
   );
 }
